Fix and Improvement
Fixed the issue when there is any slash in the file name.
Improved the Book class so it does not perform unnecessary getting process. Previously it would still get the files if one file exists, which creates unnecessary internet traffic.
Slightly modified the scrape method so that it checks the file type when there is only one button on the web page. This is only for consistency consideration, where Springer could provide only ePub but no PDF. This does not really happen though.